2.0.2_XH_alpha 20051 预支付prepayDuration改为最大时长，可以编辑，可以加减、 首页定位改为未获取到时，去上次获取的到的值 1.9.9 20034

+++ base/src/main/java/com/kernal/demo/base/bean/Street.kt
@@ -12,7 +12,7 @@ open class Street(
     var streetName: String = "",
     var ischeck: Boolean = false,
     var isCurrent: Boolean = false,
-    var prepayDuration: Int = 1
+    var prepayDuration: Double = 1.0
 ) : RealmObject(), Parcelable {
     fun copy(): Street {
         return Street(streetNo, streetName, ischeck, isCurrent, prepayDuration)
